refactor(cache): promote the write serializer to a persistence-queue module
Moves the module-level writeChain/enqueueWrite out of the useCacheManager
god-hook into a named src/lib/persistenceQueue.ts exposing enqueuePersist().
Same shared-singleton serialization semantics, now reusable and named after
its job. Sets up removing the persist-flag threading in the next commit.
